chore(pubtier): _redirects for SSE + move SSE client to frontend/src/lib/sse.js
