+++ frontend/src/main.jsx
@@ -17,53 +17,7 @@ ReactDOM.createRoot(document.getElementById('root')).render(
 
 // Option B: module-based loader wiring for bundlers. Import relative module.
 import './bhabitLogoLoaderModule.js';
-
-// --- Minimal SSE client (inline) ---
-// NOTE: adapted to JS from the TS snippet. The payload shape is:
-// { type: 'hello'|'tick'|'update'|..., updatedAt?: number, changed?: boolean }
-function startSSE(onUpdate) {
-  let es = null;
-  let retryMs = 1000;
-  const maxMs = 30000;
-
-  const connect = () => {
-    es = new EventSource('/api/events');
-
-    es.onmessage = (ev) => {
-      try {
-        const data = JSON.parse(ev.data);
-        onUpdate(data);
-      } catch (err) {
-        // ignore parse errors
-      }
-    };
-
-    es.onerror = () => {
-      try {
-        if (es) es.close();
-      } catch (e) {}
-      es = null;
-      setTimeout(connect, retryMs);
-      retryMs = Math.min(retryMs * 2, maxMs);
-    };
-
-    es.onopen = () => {
-      retryMs = 1000;
-    };
-  };
-
-  connect();
-
-  return () => {
-    try {
-      if (es) {
-        es.close();
-        es = null;
-      }
-    } catch (e) {}
-  };
-}
-// --- End SSE client ---
+import { startSSE } from './lib/sse.js';
 
 // Listen for DO updates and lightly refresh data when needed
 try {
